+++ src/game/player/libElements.ts
@@ -2,6 +2,7 @@ import Isolated from "@player/lib/isolated";
 import Say from "@player/elements/say/Say";
 import {Top, Center, Bottom, Full} from "@player/lib/PlayerFrames";
 import {VBox, HBox} from "@player/lib/UtilComponents";
+import {usePreference} from "@player/lib/preferences";
 
 export {
     Isolated,
@@ -12,4 +13,5 @@ export {
     VBox,
     HBox,
     Full,
+    usePreference,
 };